requirements.txt
discord.py==2.3.2
colorama==0.4.6
python-dotenv==1.0.1
requests==2.31.0